Header files
C++'ified the files
Added a header file.
Tidied up the code.
